# local-shared-data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Users:Current User",
        "operationId": "users:current_user",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "401": {
            "description": "Missing token or inactive user."
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Users:Patch Current User",
        "operationId": "users:patch_current_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "401": {
            "description": "Missing token or inactive user."
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                },
                "examples": {
                  "UPDATE_USER_EMAIL_ALREADY_EXISTS": {
                    "summary": "A user with this email already exists.",
                    "value": {
                      "detail": "UPDATE_USER_EMAIL_ALREADY_EXISTS"
                    }
                  },
                  "UPDATE_USER_INVALID_PASSWORD": {
                    "summary": "Password validation failed.",
                    "value": {
                      "detail": {
                        "code": "UPDATE_USER_INVALID_PASSWORD",
                        "reason": "Password should beat least 3 characters"
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/notes/": {
      "get": {
        "tags": [
          "notes"
        ],
        "summary": "Search Notes",
        "description": "\u641c\u7d22\u548c\u7b5b\u9009\u7b14\u8bb0",
        "operationId": "search_notes",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Keyword"
            }
          },
          {
            "name": "is_new",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "title": "Is New"
            }
          },
          {
            "name": "is_changed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "title": "Is Changed"
            }
          },
          {
            "name": "is_important",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "title": "Is Important"
            }
          },
          {
            "name": "author_user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Author User Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/XhsNoteResponse"
                  },
                  "title": "Response Notes-Search Notes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/notes/{note_id}": {
      "get": {
        "tags": [
          "notes"
        ],
        "summary": "Get Note Detail",
        "description": "\u83b7\u53d6\u5355\u4e2a\u7b14\u8bb0\u8be6\u60c5",
        "operationId": "get_note_detail",
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Note Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/trigger-crawl": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Trigger Crawl Task",
        "description": "\u89e6\u53d1\u722c\u53d6\u4efb\u52a1",
        "operationId": "trigger_crawl_task",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrawlTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTriggerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Tasks",
        "description": "\u83b7\u53d6\u4efb\u52a1\u5217\u8868",
        "operationId": "get_tasks",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Task Detail",
        "description": "\u83b7\u53d6\u4efb\u52a1\u8be6\u60c5",
        "operationId": "get_task_detail",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tasks"
        ],
        "summary": "Cancel Task",
        "description": "\u53d6\u6d88\u4efb\u52a1",
        "operationId": "cancel_task",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CrawlTaskRequest": {
        "properties": {
          "task_name": {
            "type": "string",
            "title": "Task Name",
            "description": "\u4efb\u52a1\u540d\u79f0"
          },
          "keyword": {
            "type": "string",
            "title": "Keyword",
            "description": "\u641c\u7d22\u5173\u952e\u8bcd"
          },
          "target_count": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Target Count",
            "description": "\u76ee\u6807\u722c\u53d6\u6570\u91cf",
            "default": 50
          },
          "sort_type": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 0.0,
            "title": "Sort Type",
            "description": "\u6392\u5e8f\u65b9\u5f0f",
            "default": 1
          },
          "cookies": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cookies",
            "description": "\u81ea\u5b9a\u4e49cookies"
          },
          "webhook_url": {
            "type": "string",
            "title": "Webhook Url",
            "description": "\u63a5\u6536\u7ed3\u679c\u7684webhook URL"
          }
        },
        "type": "object",
        "required": [
          "task_name",
          "keyword",
          "webhook_url"
        ],
        "title": "CrawlTaskRequest",
        "description": "\u722c\u53d6\u4efb\u52a1\u8bf7\u6c42"
      },
      "ErrorModel": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorModel"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TaskTriggerResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "github_run_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Run Url"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "TaskTriggerResponse",
        "description": "\u4efb\u52a1\u89e6\u53d1\u54cd\u5e94"
      },
...
      "UserRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "is_superuser": {
            "type": "boolean",
            "title": "Is Superuser",
            "default": false
          },
          "is_verified": {
            "type": "boolean",
            "title": "Is Verified",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "email"
        ],
        "title": "UserRead"
      },
      "UserUpdate": {
        "properties": {
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "is_superuser": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Superuser"
          },
          "is_verified": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Verified"
          }
        },
        "type": "object",
        "title": "UserUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "XhsNoteResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "note_id": {
            "type": "string",
            "title": "Note Id"
          },
          "note_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Url"
          },
          "note_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Type"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "desc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Desc"
          },
          "author_nickname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author Nickname"
          },
          "liked_count": {
            "type": "integer",
            "title": "Liked Count"
          },
          "comment_count": {
            "type": "integer",
            "title": "Comment Count"
          },
          "current_tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Tags"
          },
          "is_new": {
            "type": "boolean",
            "title": "Is New"
          },
          "is_changed": {
            "type": "boolean",
            "title": "Is Changed"
          },
          "is_important": {
            "type": "boolean",
            "title": "Is Important"
          },
          "first_crawl_time": {
            "type": "string",
            "format": "date-time",
            "title": "First Crawl Time"
          },
          "last_crawl_time": {
            "type": "string",
            "format": "date-time",
            "title": "Last Crawl Time"
          }
        },
        "type": "object",
        "required": [
          "id",
          "note_id",
          "note_url",
          "note_type",
          "title",
          "desc",
          "author_nickname",
          "liked_count",
          "comment_count",
          "current_tags",
          "is_new",
          "is_changed",
          "is_important",
          "first_crawl_time",
          "last_crawl_time"
        ],
        "title": "XhsNoteResponse",
        "description": "\u7b14\u8bb0\u54cd\u5e94\u6a21\u5f0f"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/jwt/login"
          }
        }
      }
    }
  }
}